requirements.txt
customtkinter==5.2.1
reportlab==4.0.7
Pillow>=10.0.0
